+++ src/integrations/AltTextAi.php
@@ -2,6 +2,7 @@
 
 namespace GlueAgency\ImageCaption\integrations;
 
+use Craft;
 use craft\elements\Asset;
 use GlueAgency\ImageCaption\ImageCaption;
 use GlueAgency\ImageCaption\integrations\responses\ErrorResponse;
@@ -16,6 +17,7 @@ class AltTextAi extends AbstractIntegration implements IntegrationInterface
 
     const TOKEN = 'token';
     const ENDPOINT = 'endpoint';
+    const KEYWORDS = 'keywords';
 
     public string $name = 'Alt Text Ai';
 
@@ -25,14 +27,24 @@ class AltTextAi extends AbstractIntegration implements IntegrationInterface
             $client = $this->getAuthorizedClient();
             $language = strtok($asset->getSite()->language, '-');
 
+            $params = [
+                'lang'  => $language,
+                'image' => [
+                    'asset_id' => $asset->id,
+                    'url'      => $asset->getUrl(),
+                ],
+            ];
+
+            if($keywords = $this->getSetting(self::KEYWORDS))
+            {
+                $parts = explode(' ', $keywords);
+                $allowed = array_slice($parts, 0, 6);
+
+                $params['keywords'] = $allowed;
+            }
+
             $response = $client->post('api/v1/images', [
-                'json' => [
-                    'lang'  => $language,
-                    'image' => [
-                        'asset_id' => $asset->id,
-                        'url'      => $asset->getUrl(),
-                    ],
-                ]
+                'json' => $params,
             ]);
 
             $data = json_decode($response->getBody()->getContents(), true);
@@ -50,7 +62,7 @@ class AltTextAi extends AbstractIntegration implements IntegrationInterface
         return [
             new IntegrationSetting(
                 IntegrationSetting::AUTOCOMPLETE,
-                'Endpoint',
+                Craft::t('image-captions', 'Endpoint'),
                 self::ENDPOINT,
                 null,
                 true,
@@ -58,15 +70,21 @@ class AltTextAi extends AbstractIntegration implements IntegrationInterface
             ),
             new IntegrationSetting(
                 IntegrationSetting::PASSWORD,
-                'Token',
+                Craft::t('image-captions','Token'),
                 self::TOKEN,
                 null,
                 true
+            ),
+            new IntegrationSetting(
+                IntegrationSetting::TEXT,
+                Craft::t('image-captions','Keywords'),
+                self::KEYWORDS,
+                Craft::t('image-captions','Add up to 6 English keywords separated by a space that should be considered when generating the alt text.'),
             )
         ];
     }
 
-    public function getAuthorizedClient(): Client
+    protected function getAuthorizedClient(): Client
     {
         return new Client([
             'base_uri' => $this->getSetting(self::ENDPOINT),