+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Image Captions
 
+## 4.0.2 - 2023-06-29
+
+- Add `keywords` to the AltTextAi provider
+- Fix Providers unique `handle` validation
+
 ## 4.0.1 - 2023-06-29
 
 - Fix Project config apply for Configurations